pyproject.toml
[tool.poetry]
name = "p1"
version = "0.1.0"
description = ""
authors = ["gglo0ol <124757163+gglo0ol@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
uvicorn = "^0.28.0"
pydantic-settings = "^2.2.1"
black = "^24.3.0"
asyncpg = "^0.29.0"
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
